backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
channels==4.1.0
daphne==4.1.2
psycopg[binary]==3.2.3
PyJWT==2.9.0
